refactor(v2): extract moderation HTTP handlers into moderation.rs
Move the moderation route handlers, auth helpers, and moderation-only
request/response types out of main.rs (612 lines) into the existing
moderation module, following the turns.rs extraction model. Pure code
motion: no behavior changes, all 393 orchestrator tests pass.

Also aligns the orchestrator crate version with the root release
version (0.0.87), which the content registry test enforces.
